% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_el.R
\name{append_el}
\alias{append_el}
\title{Append value(s) to a named list element}
\usage{
append_el(x, lst, targ)
}
\arguments{
\item{x}{Values to append}

\item{lst}{The list to append to. Must be named and contain the name as denoted by \emph{targ}.}

\item{targ}{The element in \emph{lst} to append to}
}
\value{
The input list, with values within \emph{val} appended to \emph{targ}.
}
\description{
Append a vector of values to a specific named element in a list.
}
\details{
This is simply a convenient way to append a vector of values to a named list element. It is
convenient in the sense that instead of:

\code{lst[["targ"]] <- c(lst[["targ"]], x)},

you can do:

\code{lst <- append_el(x, lst = lst, targ = "targ")}

The main use case is e.g. appending one or more segment(ids) to a base argument list for
comparison and/or testing.
}
\examples{
# TBD
}
