% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_monthChunks.R
\name{make_monthChunks}
\alias{make_monthChunks}
\title{Make sequences of start and end months based on ymd input}
\usage{
make_monthChunks(date_start, date_end)
}
\arguments{
\item{date_start}{A vector of length 1 denoting the start date range. Must be in ymd-compatible format}

\item{date_end}{A vector of length 1 denoting the end date range. Must be in ymd-compatible format}
}
\value{
A named \code{list} of length 2, with as many elements as months based on the input start and end. One
list contains start dates, and the other end dates.

All elements are returned as class \code{Date}.
}
\description{
For a start and end date, formatted in 'ymd'-compatible format, create sequences
}
\examples{
make_monthChunks("2017-01-01", "2017-03-01")
make_monthChunks(20170101, 20170301)
make_monthChunks("2017-Jan-1", 20170301)
make_monthChunks("2017-January 1", 20170301)
}
