% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_actName.R
\name{normalize_actName}
\alias{normalize_actName}
\title{String normalization for common account name formats}
\usage{
normalize_actName(x, strip_prefix = TRUE, fix_markup = TRUE,
  fix_encoding = TRUE)
}
\arguments{
\item{x}{(required) A character vector to normalize}

\item{strip_prefix}{(optional) Should all characters up to and including \code{:} be stripped?
Defaults to \code{TRUE}}

\item{fix_markup}{(optional) Should html-markup be escaped? Defaults to \code{TRUE}; and
calls \code{\link[wzMisc]{unescape_markup}} with fixed parameters; see details.}

\item{fix_encoding}{(optional) Should encoding be normalized? Defaults to \code{TRUE} and
calls \code{\link[base]{iconv}} with fixed parameters; see details}
}
\value{
A normalized vector
}
\description{
Strip out certain specials, make lowercase, optionally parse and handle markup and encoding
}
\details{
TBD
}
\examples{
# TBD
}
