% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_query_ret.R
\name{process_query_ret}
\alias{process_query_ret}
\title{Clean up RSiteCatalyst Queue* query returns}
\usage{
process_query_ret(x)
}
\arguments{
\item{x}{A call return from an \code{RSiteCatalyst} Queue* function. Expected to be a
\code{data.frame}.}
}
\value{
A \code{data.table}, with any \code{POSIXlt} or \code{POSIXct} columns coerced to the \code{data.table}
\code{IDate} class. Additionally, any columns where all values are blank are dropped, with a message.
}
\description{
For a query return, transform to data.table, handle POSIXlt, and drop columns that are all blank
}
\examples{
#TBD
}
