% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enqueue_batch.R
\name{enqueue_batch}
\alias{enqueue_batch}
\title{Batch enqueue and pull RSiteCatalyst Queue* reports}
\usage{
enqueue_batch(argList, fun = "QueueRanked", use_names = TRUE,
  auto_enqueue = TRUE, enqueue_only = FALSE, ...)
}
\arguments{
\item{argList}{(required) A (nested) list of named arguments to pass to \emph{fun}}

\item{fun}{(optional) A \code{quoted} character vector of length 1. Defaults to
\link[RSiteCatalyst]{QueueRanked}.}

\item{use_names}{(optional) A logical vector of length 1. Should the names of
\emph{argList} be automatically appended to the output result? Defaults to
\code{TRUE} and igored in-function if no or missing names detected}

\item{auto_enqueue}{(optional) A logical vector of length 1. Should we append an
argument of \code{enqueueOnly} and set the value to \code{TRUE} if this argument
is not detected in all element-lists of \emph{argList}? Defaults to \code{TRUE}.}

\item{enqueue_only}{(optional) A logical vector of length 1. Only enqueue, i.e.
do not attempt to also pull the report? Defaults to \code{FALSE}.}

\item{...}{(optional) Additional arguments to pass to \link[RSiteCatalyst]{GetReport}}
}
\value{
A list the same length of \emph{argList} containing either the \code{data.frame} output from
\code{GetReport} or the (\code{integer}) report ID. The latter applies when an error is
encountered while trying to pull the report, which almost always happens because
one or more submitted reports requires additional processing time.

If \code{use_names = TRUE} and all names within \emph{argList} are
present, the output elements will be named using the corresponding (positional) names
of \emph{argList}.

Each \code{data.frame} element will have an attribute called \code{reportID}
to denote the ID from each respective enqueued report, in the event
you wish to e.g. re-pull one or more reports. \code{integer} elements (i.e. report IDs only)
will not contain attributes, since this would be redundant (and potentially confusing).
}
\description{
For a list of arguments in the scope of a Queue* function, enqueue and pull data
}
\details{
The list of valid arguments for \emph{fun} includes all \code{RSiteCatalyst}
Queue\code{*} functions where \emph{enqueueOnly} is a valid argument. This
currently includes:

\itemize{
\item QueueRanked
\item QueueTrended
\item QueuePathing
\item QueueFallout
\item QueueOvertime
\item QueueSummary
\item QueueDataWarehouse
}
}
\examples{
#TBD
}
